+++ app/src/main/java/com/victor/dogbreeds/ui/breedDetails/BreedDetailsContract.kt
@@ -1,4 +1,4 @@
-package com.victor.dogbreeds.breedDetails
+package com.victor.dogbreeds.ui.breedDetails
 
 interface BreedDetailsContract {
     interface View {
