+++ app/src/main/java/com/victor/dogbreeds/ui/editProfile/EditProfileContract.kt
@@ -1,4 +1,4 @@
-package com.victor.dogbreeds.editProfile
+package com.victor.dogbreeds.ui.editProfile
 
 interface EditProfileContract {
     interface View{
