+++ app/src/main/java/com/victor/dogbreeds/ui/home/HomeContract.kt
@@ -1,4 +1,4 @@
-package com.victor.dogbreeds.home
+package com.victor.dogbreeds.ui.home
 
 interface HomeContract {
     interface View {
