+++ app/src/main/java/com/victor/dogbreeds/ui/signIn/SignInActivity.kt
@@ -1,11 +1,11 @@
-package com.victor.dogbreeds.signIn
+package com.victor.dogbreeds.ui.signIn
 
 import android.content.Context
 import android.content.Intent
 import com.victor.dogbreeds.R
-import com.victor.dogbreeds.base.BaseActivity
-import com.victor.dogbreeds.home.HomeActivity
-import com.victor.dogbreeds.signUp.SignUpActivity
+import com.victor.dogbreeds.ui.base.BaseActivity
+import com.victor.dogbreeds.ui.home.HomeActivity
+import com.victor.dogbreeds.ui.signUp.SignUpActivity
 import kotlinx.android.synthetic.main.activity_sign_in.*
 import org.koin.android.ext.android.inject
 import org.koin.core.parameter.parametersOf
